+++ examples/swift-interop/src/swift/swiftGreeter/Greeter.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+@objcMembers public class Greeter: NSObject {
+    @objc(greetWithName:)
+    public static func greet(name: String) -> String {
+        return "Hello \(name), from Swift!"
+    }
+}